Cadastro de Empresas

+++ src/Extra.hs
@@ -3,9 +3,9 @@
 module Extra where
 
 import Database.Persist.TH
-
-data Perfil = PadrinhoPerfil {padrinhoPerfil :: Int } |
-              ResponsavelPerfil { responsavelPerfil :: Int } |
-              EmpresaPerfil { empresaPerfil :: Int }
+import Data.Int
+data Perfil = PadrinhoPerfil {padrinhoPerfil :: Int64 } |
+              ResponsavelPerfil { responsavelPerfil :: Int64 } |
+              EmpresaPerfil { empresaPerfil :: Int64 }
               deriving (Show, Read, Eq)
 derivePersistField "Perfil"
